# openapi.yaml
openapi: 3.0.3
info:
  title: API Sổ thu chi cá nhân
  version: 2.0.0
  description: REST API cho TASK_00030. Mọi response JSON sử dụng envelope success/message/data hoặc success/error.
servers:
  - url: http://localhost:3000/api/v1
...
paths:
...
  /budgets:
    get:
      tags: [Budgets]
      summary: Danh sách ngân sách kèm tiến độ
      responses: { '200': { $ref: '#/components/responses/Success' } }
    post:
      tags: [Budgets]
      summary: Tạo ngân sách
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BudgetInput' } } }
      responses: { '201': { $ref: '#/components/responses/Success' } }
  /budgets/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Budgets]
      summary: Chi tiết ngân sách
      responses: { '200': { $ref: '#/components/responses/Success' } }
    patch:
      tags: [Budgets]
      summary: Sửa ngân sách
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BudgetInput' } } }
      responses: { '200': { $ref: '#/components/responses/Success' } }
    delete:
      tags: [Budgets]
      summary: Xóa ngân sách
      responses: { '200': { $ref: '#/components/responses/Success' } }
  /goals:
    get:
      tags: [Goals]
      summary: Danh sách mục tiêu
      responses: { '200': { $ref: '#/components/responses/Success' } }
    post:
      tags: [Goals]
      summary: Tạo mục tiêu
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/GoalInput' } } }
      responses: { '201': { $ref: '#/components/responses/Success' } }
  /goals/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Goals]
      summary: Chi tiết mục tiêu và lịch sử đóng góp
      responses: { '200': { $ref: '#/components/responses/Success' } }
    patch:
      tags: [Goals]
      summary: Sửa mục tiêu
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/GoalInput' } } }
      responses: { '200': { $ref: '#/components/responses/Success' } }
    delete:
      tags: [Goals]
      summary: Xóa mục tiêu
      responses: { '200': { $ref: '#/components/responses/Success' } }
  /goals/{id}/contributions:
    post:
      tags: [Goals]
      summary: Ghi tăng/giảm tiến độ mục tiêu
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: number, example: 1000000 }
                note: { type: string }
      responses: { '201': { $ref: '#/components/responses/Success' } }
...
  /goals/{id}/pause:
    post:
      tags: [Goals]
      summary: Tạm dừng mục tiêu
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { $ref: '#/components/responses/Success' } }
  /goals/{id}/resume:
    post:
      tags: [Goals]
      summary: Tiếp tục mục tiêu
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { $ref: '#/components/responses/Success' } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { in: path, name: id, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Success:
      description: Thành công
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean, example: true }
        message: { type: string, example: Thành công. }
        data: { nullable: true }
        meta: { type: object, nullable: true }
...
    BudgetInput:
      type: object
      required: [name, amount, startDate, endDate]
      properties:
        name: { type: string, example: Ăn uống tháng 9 }
        categoryId: { type: string, format: uuid, nullable: true }
        amount: { type: number, example: 3000000 }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
    GoalInput:
      type: object
      required: [name, targetAmount]
      properties:
        name: { type: string, example: Quỹ dự phòng }
        targetAmount: { type: number, example: 50000000 }
        walletId: { type: string, format: uuid, nullable: true }
        targetDate: { type: string, format: date, nullable: true }
        status: { type: string, enum: [ACTIVE, COMPLETED, CANCELLED] }
security:
  - bearerAuth: []